--- apps/back-office/src/components/manage-service/hooks/index.ts
@@ -1,3 +1,0 @@
-import useManagePagination from "./use-manage-pagination";
-
-export { useManagePagination };
