+++ apps/back-office/src/components/manage-service/manage-service.meta.ts
@@ -16,11 +16,3 @@ export const getPhraseItemListMocks = (num: number): Array<PhraseItemWithId> =>
     ...phraseItemMocks,
     phraseId: idx + 1,
   }));
-
-export const rowsPerPageOptions = [
-  { value: "10", label: "10개" },
-  { value: "20", label: "20개" },
-  { value: "30", label: "30개" },
-  { value: "50", label: "50개" },
-  { value: "100", label: "100개" },
-];
