--- apps/back-office/src/components/manage-service/manage-service.type.ts
@@ -1,9 +1,0 @@
-import { Checkbox } from "../ui/checkbox";
-
-type OnCheck = Parameters<typeof Checkbox>[0]["onCheckedChange"];
-type OnDelete = (
-  id: number,
-  ...params: Parameters<NonNullable<OnCheck>>
-) => void;
-
-export type { OnCheck, OnDelete };
